+++ locales/en/quiz.json
@@ -152,10 +152,13 @@
         "hover": "Click to cancel",
         "notHover": "To what degree? "
       },
+      "oneClickAnswering": "One-click answering",
       "answer": {
         "neutral": "I don't know",
+        "stronglyAgree" : "Strongly agree",
         "agree": "Agree",
-        "disagree": "Disagree"
+        "disagree": "Disagree",
+        "stronglyDisagree": "Strongly disagree"
       },
       "weight": {
         "1": "Somewhat",
