+++ src/components/Survey/SingleSurveyPage/SurveyAnswers/MultiChoiceAnswers/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./MultiChoiceAnswersView";
