+++ src/components/Survey/SingleSurveyPage/SurveyAnswers/SingleChoiceAnswers/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./SingleChoiceAnswersView";
