+++ src/components/Survey/SingleSurveyPage/SurveyAnswers/SurveyAnswersView.tsx
@@ -1,16 +1,15 @@
-import React from "react";
-import { SurveyAnswerType } from "@generated/graphql";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { faCheck, faTimes } from "@fortawesome/free-solid-svg-icons";
-import Button from "@shared/Button";
-import { CSSTransition } from "react-transition-group";
+import React, { useState } from "react";
 import {
   SlimAnswer,
   UseSurveyActions,
   UseSurveyData,
 } from "@components/Survey/utils/useSurvey";
+import { Container } from "./SurveyAnswersStyle";
+import MultiChoiceAnswersView from "@components/Survey/SingleSurveyPage/SurveyAnswers/MultiChoiceAnswers";
+import SingleChoiceAnswersView from "@components/Survey/SingleSurveyPage/SurveyAnswers/SingleChoiceAnswers";
+import { CSSTransition } from "react-transition-group";
 import useTranslation from "next-translate/useTranslation";
-import { AnswerButton, Answers, Container } from "./SurveyAnswersStyle";
+import { AnsweringModeToggle } from "./SurveyAnswersStyle";
 
 interface Props {
   actions: UseSurveyActions;
@@ -19,124 +18,39 @@ interface Props {
 
 const SurveyAnswers: React.FC<Props> = ({ actions, data }) => {
   const { t } = useTranslation("quiz");
+
+  const [isOneClick, setIsOneClick] = useState(false);
   const { nextQuestion, viewResults } = actions;
-  const { currentQuestion } = data;
-  const [tempAnswer, setTempAnswer] = React.useState<SlimAnswer | undefined>();
-  const [isHover, setHover] = React.useState<boolean>(false);
+  const { currentQuestion, questions } = data;
 
   const pushAnswer = async (answer: SlimAnswer) => {
-    if (await nextQuestion(answer)) {
-      setTempAnswer(undefined);
-    } else {
+    if (!(await nextQuestion(answer))) {
       await viewResults();
     }
   };
 
-  const handleMainAnswer = (type: SurveyAnswerType) => {
-    const answer = {
-      type,
-      weight: 0,
-      question: {
-        id: currentQuestion.id,
-      },
-    };
-
-    if (type === SurveyAnswerType.Neutral) {
-      pushAnswer(answer);
-      return;
-    }
-
-    setTempAnswer(answer);
-  };
-
-  const handleWeightAnswer = (weight: number) => {
-    const answer = {
-      ...tempAnswer,
-      weight,
-    };
-
-    pushAnswer(answer);
-  };
-
-  const handleResetWeightAnswer = () => {
-    setTempAnswer(undefined);
-  };
+  const isPastFifthQuestion = questions.indexOf(currentQuestion) > 4;
 
   return (
     <Container>
-      {tempAnswer && (
-        <Answers>
-          <AnswerButton
-            onClick={handleResetWeightAnswer}
-            variant={tempAnswer.type}
-            answer={tempAnswer}
-            onMouseEnter={() => setHover(true)}
-            onMouseLeave={() => setHover(false)}
-          >
-            <span>
-              <FontAwesomeIcon
-                icon={
-                  tempAnswer.type === SurveyAnswerType.Agree ? faCheck : faTimes
-                }
-              />
-            </span>
-            <span>
-              {isHover
-                ? t("survey.actions.main.hover")
-                : t("survey.actions.main.notHover")}
-            </span>
-          </AnswerButton>
-        </Answers>
+      {!isPastFifthQuestion && (
+        <AnsweringModeToggle
+          value={isOneClick}
+          onValueChange={setIsOneClick}
+          label={t("survey.actions.oneClickAnswering")}
+        />
       )}
-      {!tempAnswer && (
-        <Answers>
-          <AnswerButton
-            variant={SurveyAnswerType.Disagree}
-            answer={tempAnswer}
-            onClick={() => handleMainAnswer(SurveyAnswerType.Disagree)}
-          >
-            <div>
-              <span>
-                <FontAwesomeIcon icon={faTimes} />
-              </span>
-              <span>{t("survey.actions.answer.disagree")}</span>
-            </div>
-          </AnswerButton>
-          <AnswerButton
-            variant={SurveyAnswerType.Neutral}
-            answer={tempAnswer}
-            onClick={() => handleMainAnswer(SurveyAnswerType.Neutral)}
-          >
-            <span>{t("survey.actions.answer.neutral")}</span>
-          </AnswerButton>
-          <AnswerButton
-            variant={SurveyAnswerType.Agree}
-            answer={tempAnswer}
-            onClick={() => handleMainAnswer(SurveyAnswerType.Agree)}
-          >
-            <span>
-              <FontAwesomeIcon icon={faCheck} />
-            </span>
-            <span>{t("survey.actions.answer.agree")}</span>
-          </AnswerButton>
-        </Answers>
-      )}
-      <CSSTransition
-        in={tempAnswer}
-        classNames="fade"
-        unmountOnExit
-        addEndListener={(node, done) => {
-          node.addEventListener("transitionend", done, false);
-        }}
-      >
-        <Answers>
-          <Button isFullWidth onClick={() => handleWeightAnswer(1)}>
-            {t("survey.actions.weight.1")}
-          </Button>
-          <Button isFullWidth onClick={() => handleWeightAnswer(2)}>
-            {t("survey.actions.weight.2")}
-          </Button>
-        </Answers>
+      <CSSTransition in={isOneClick} classNames="fade" unmountOnExit>
+        <SingleChoiceAnswersView
+          questionId={currentQuestion.id}
+          onAnswer={pushAnswer}
+        />
+      </CSSTransition>
+      <CSSTransition in={!isOneClick} classNames="fade" unmountOnExit>
+        <MultiChoiceAnswersView
+          questionId={currentQuestion.id}
+          onAnswer={pushAnswer}
+        />
       </CSSTransition>
     </Container>
   );
