+++ src/components/Survey/SingleSurveyPage/SurveyAnswers/types.ts
@@ -1,0 +1,6 @@
+import { SlimAnswer } from "@components/Survey/utils/useSurvey";
+
+export interface BaseAnswersViewProps {
+  questionId: string;
+  onAnswer(answer: SlimAnswer);
+}
